requirements.txt
dj-database-url==0.5.0
Django==2.2.5
gunicorn==19.9.0
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
psycopg2-binary==2.7.7
Pillow==6.1.0
djangorestframework==3.10.3